Move some SCDB related functions out of validation
* Refactor: move SCDB related functions out of validation and into
SidechainDB.
